requirements.txt
Django==3.1.2
djangorestframework==3.12.1
Markdown==3.2.2
requests==2.24.0